order details backend

+++ backend/orderDetails.php
@@ -5,7 +5,7 @@
     if(isset($_POST['orderID'])){
 
         $orderID = (int) mysqli_real_escape_string($conn, $_POST['orderID']);
-        $status = (int) mysqli_real_escape_string($conn, $_POST['status']);
+        $status = mysqli_real_escape_string($conn, $_POST['status']);
 
         $sql1 = mysqli_query($conn, "UPDATE orderMaster SET status = '{$status}' WHERE id = '{$orderID}'");
         if($sql1){
